+++ app/views/scripts/index.phtml
