--- app/views/scripts/json/add.phtml
@@ -1,102 +1,0 @@
-<html>
-<head>
-  <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport"    content="width=device-width, initial-scale=1.0">
-    <meta name="author"      content="Ignasi Ortiz, Carlos Leichner, Gabriel Farrus" />
-	<meta name="description" content="IT Academy - FullStack LAMP" />
-    <title>PhP Developers</title>
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body>
-    
-    <nav method = "$POST" class="flex justify-center space-x-4">
-        
-        
-        <a href="/views/scripts/index.html" class="font-bold px-3 py-2 text-slate-700 rounded-lg hover:bg-slate-100 hover:text-slate-900">Create Task</a>
-        
-      </nav>
-    
-    
-      <form class="flex justify-center p-10">
-        <label class="block">
-
-          <span class=" block text-sm font-medium text-slate-700">ID Task:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST id" name="id_Task" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-          <span class=" block text-sm font-medium text-slate-700">Task Description:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST description" name ="description" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-          <span class=" block text-sm font-medium text-slate-700">Creator User:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST master" name="masterUsr_id" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-          <span class=" block text-sm font-medium text-slate-700">User:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST slave" name="slaveUsr_id" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-          <span class=" block text-sm font-medium text-slate-700">Expedition Date:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST expedition date" name="created_At" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-          <span class=" block text-sm font-medium text-slate-700">Start Date:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST initiated" name="initiated" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-          <span class=" block text-sm font-medium text-slate-700">Finish Date:</span>
-          <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="$_POST done" name="done" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
-            focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
-            disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
-            invalid:border-pink-500 invalid:text-pink-600
-            focus:invalid:border-pink-500 focus:invalid:ring-pink-500
-          "/>
-
-        </label>
-        <!-- ... -->
-    </form>
-      
-    <madre method = "$POST" class="flex justify-center space-x-4 ">
-
-        <button class=" bg-sky-500 px-3 hover:bg-green-700 ...">
-         <a href="/addtask">Continue</a>
-        </button>
-
-       <button class=" bg-sky-500 px-3 hover:bg-red-700 ...">
-        <a href="/views/scripts/index.html">Cancel</a>
-       </button>
-
-   </madre>
-   <footer class="mt-1  px-3 py-2">
-    <p class="font-extrabold text-2xl bg-blue-400 text-center"><b>IT Academy (itinerari LAMP) - Cibernàrium de BcnActiva - 2022</b></p>
-  </footer>
-
-   </html>
-      
-
