+++ app/views/scripts/task/del.phtml
@@ -10,19 +10,19 @@
 </head>
 <body>
     
-    <nav method = "POST" class="flex justify-center space-x-4">
+    <nav method = "$POST" class="flex justify-center space-x-4">
         
         
-        <a href="/views/scripts/index.html" class="font-bold px-3 py-2 text-slate-700 rounded-lg hover:bg-slate-100 hover:text-slate-900">Search Task</a>
+        <a href="/views/scripts/index.phtml" class="font-bold px-3 py-2 text-slate-700 rounded-lg hover:bg-slate-100 hover:text-slate-900">Delete Task</a>
         
       </nav>
     
     
-      <form method = "POST" class="flex justify-center p-10">
+      <form class="flex justify-center p-10">
         <label class="block">
           <span class=" block text-sm font-medium text-slate-700">ID Task:</span>
           <!-- Using form state modifers, the classes can be identical for every input -->
-          <input type="text" value="" name ="id_task" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
+          <input type="text" value="$_POST id" name ="id_Task" class="mt-1  px-3 py-2 bg-white border border-slate-300 rounded-md text-sm shadow-sm placeholder-slate-400
             focus:outline-none focus:border-sky-500 focus:ring-1 focus:ring-sky-500
             disabled:bg-slate-50 disabled:text-slate-500 disabled:border-slate-200 disabled:shadow-none
             invalid:border-pink-500 invalid:text-pink-600
@@ -37,11 +37,11 @@
     <madre method = "POST" class="flex justify-center space-x-4 ">
 
         <button class=" bg-sky-500 px-3 hover:bg-green-700 ...">
-         <a href="/edittask">Continue</a>
+         <a href="task/deltask">Continue</a>
         </button>
 
        <button class=" bg-sky-500 px-3 hover:bg-red-700 ...">
-        <a href="/views/scripts/index.html">Cancel</a>
+        <a href="scripts/index.phtml">Cancel</a>
        </button>
 
    </madre>
