+++ app/views/scripts/task/edit.phtml
@@ -65,7 +65,7 @@
         </button>
 
        <button class=" bg-sky-500 px-3 hover:bg-red-700 ...">
-        <a href="/views/scripts/index.html">Cancel</a>
+        <a href="scripts/index.phtml">Cancel</a>
        </button>
 
    </madre>
