+++ app/views/scripts/task/index.phtml
@@ -93,11 +93,11 @@
       <madre method = "POST" class="flex justify-center space-x-4 ">
   
           <button class=" bg-sky-500 px-3 hover:bg-green-700 ...">
-           <a href="/listtask">Continue</a>
+           <a href="../web/listask">Continue</a>
           </button>
   
          <button class=" bg-sky-500 px-3 hover:bg-red-700 ...">
-          <a href="/views/scripts/index.html">Cancel</a>
+          <a href="../web/listask">Cancel</a>
          </button>
   
      </madre>
